+++ portal-impl/src/com/liferay/portal/upgrade/v7_0_3/UpgradeSybase.java
@@ -30,15 +30,17 @@ import java.sql.SQLException;
 /**
  * @author Cristina González
  */
-public class UpgradeNVarcharSQLServer extends UpgradeProcess {
+public class UpgradeSybase extends UpgradeProcess {
 
-	protected void alterNVarcharColumns() throws Exception {
-		StringBundler sb = new StringBundler(4);
+	protected void alterVarcharColumns() throws Exception {
+		StringBundler sb = new StringBundler(6);
 
-		sb.append("select table_name, column_name from ");
-		sb.append("INFORMATION_SCHEMA.COLUMNS ");
-		sb.append("where DATA_TYPE = 'nvarchar' ");
-		sb.append("and character_maximum_length = 2000");
+		sb.append("select o.name, c.name ");
+		sb.append("from sysobjects o, syscolumns c, systypes t ");
+		sb.append("where o.id = c.id ");
+		sb.append("and c.type = t.type ");
+		sb.append("and t.name = 'varchar' ");
+		sb.append("and c.length = 1000");
 
 		try (LoggingTimer loggingTimer = new LoggingTimer();
 			PreparedStatement ps = connection.prepareStatement(sb.toString());
@@ -46,12 +48,13 @@ public class UpgradeNVarcharSQLServer extends UpgradeProcess {
 
 			while (rs.next()) {
 				String tableName = rs.getString(1);
+
 				String columnName = rs.getString(2);
 
 				try {
 					runSQL(
-						"alter table " + tableName + " alter column " +
-							columnName + " nvarchar(4000)");
+						"alter table " + tableName + " modify " + columnName +
+							" varchar(4000)");
 				}
 				catch (SQLException sqle) {
 					if (sqle.getErrorCode() == 1441) {
@@ -80,14 +83,13 @@ public class UpgradeNVarcharSQLServer extends UpgradeProcess {
 	protected void doUpgrade() throws Exception {
 		DB db = DBManagerUtil.getDB();
 
-		if (db.getDBType() != DBType.SQLSERVER) {
+		if (db.getDBType() != DBType.SYBASE) {
 			return;
 		}
 
-		alterNVarcharColumns();
+		alterVarcharColumns();
 	}
 
-	private static final Log _log = LogFactoryUtil.getLog(
-		UpgradeNVarcharSQLServer.class);
+	private static final Log _log = LogFactoryUtil.getLog(UpgradeSybase.class);
 
 }
